Code Update
Added an animation in the cstr.jl, however I couldn't figure it out how to add a suptitlte (like in matplotlib) in order to show the current coolant temperature in the animation. I've also added Measures to the project to change the left_margin in the subplots for the animation.